Create NOTES - NO PAYMENT MONTH THREE.vbs
This is the script to follow when no payment is made on account for three months.  This document is about 80% complete.  There may need to be additional looping to add.  I have run this and it appears to run properly but I would am going to go back in and run it again to be sure everything is completing properly.